requirements.txt
python-decouple==3.6
pytz==2021.3
requests==2.27.1
timezonefinder==5.2.0